requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
django-filter==22.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
phonenumbers==8.13.6
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.14.0
